requirements.txt
urlextract
wordcloud
pandas
emoji
streamlit
matplotlib
seaborn
plotly
pillow